+++ src-ui/app/app.module.ts
@@ -154,6 +154,11 @@ import { ObfuscatedValueDirective } from './directives/obfuscated-value.directiv
 import { HeartRateCalmPeriodEnableSleepModeModalComponent } from './views/dashboard-view/views/sleep-detection-view/heart-rate-calm-period-enable-sleepmode-modal/heart-rate-calm-period-enable-sleep-mode-modal.component';
 import { SleepModeEnableOnHeartRateCalmPeriodAutomationService } from './services/sleep-detection-automations/sleep-mode-enable-on-heart-rate-calm-period-automation.service';
 import { HeartRateChartComponent } from './views/dashboard-view/views/sleep-detection-view/heart-rate-calm-period-enable-sleepmode-modal/heart-rate-chart/heart-rate-chart.component';
+import { StartWithSteamVRHowToModalComponent } from './views/dashboard-view/views/settings-view/settings-general-tab/confirm-modal/start-with-steamvr-how-to-modal.component';
+import { QuitWithSteamVRService } from './services/quit-with-steamvr.service';
+import { VRChatMicMuteAutomationService } from './services/osc-automations/vrchat-mic-mute-automation.service';
+import { MiscTestingComponent } from './components/developer-debug-modal/misc-testing/misc-testing.component';
+import { VRChatMicMuteAutomationsViewComponent } from './views/dashboard-view/views/vrchat-mic-mute-automations-view/vrchat-mic-mute-automations-view.component';
 
 [localeEN, localeFR, localeCN_TW, localeNL, localeKO, localeJP, localeES, localeID].forEach(
   (locale) => registerLocaleData(locale)
@@ -243,6 +248,9 @@ export function createTranslateLoader(http: HttpClient) {
     ObfuscatedValueDirective,
     HeartRateCalmPeriodEnableSleepModeModalComponent,
     HeartRateChartComponent,
+    StartWithSteamVRHowToModalComponent,
+    MiscTestingComponent,
+    VRChatMicMuteAutomationsViewComponent,
   ],
   imports: [
     CommonModule,
@@ -300,6 +308,7 @@ export class AppModule {
     private deepLinkService: DeepLinkService,
     private sleepPreparationService: SleepPreparationService,
     private pulsoidService: PulsoidService,
+    private quitWithSteamVRService: QuitWithSteamVRService,
     // GPU automations
     private gpuAutomations: GpuAutomationsService,
     // Sleep mode automations
@@ -320,6 +329,7 @@ export class AppModule {
     // OSC automations
     private oscGeneralAutomationsService: OscGeneralAutomationsService,
     private sleepingAnimationsAutomationService: SleepingAnimationsAutomationService,
+    private vrchatMicMuteAutomationService: VRChatMicMuteAutomationService,
     // Status automations
     private statusChangeForPlayerCountAutomationService: StatusChangeForPlayerCountAutomationService,
     // Invite automations
@@ -369,6 +379,7 @@ export class AppModule {
           this.notificationService.init(),
           this.sleepPreparationService.init(),
           this.pulsoidService.init(),
+          this.quitWithSteamVRService.init(),
         ]);
         // Initialize GPU control services
         await this.sidecarService.init().then(async () => {
@@ -407,6 +418,7 @@ export class AppModule {
           // OSC automations
           this.oscGeneralAutomationsService.init(),
           this.sleepingAnimationsAutomationService.init(),
+          this.vrchatMicMuteAutomationService.init(),
           // Status automations
           this.statusChangeForPlayerCountAutomationService.init(),
           // Invite automations